@@ -8,30 +8,33 @@ class CitiesControllerTest < ActionController::TestCase
   end
 
   test "should get index" do
-    get :index, locale: :en
+    get :index, params: { locale: :en }
     assert_response :success
     assert_not_nil assigns(:cities)
   end
 
   test "unauthenticated should not get new" do
-    get :new, locale: :en
+    get :new, params: { locale: :en }
     assert_response :redirect
   end
 
   test "authenticated should get new" do
     sign_in users(:admin)
-    get :new, locale: :en
+    get :new, params: { locale: :en }
     assert_response :success
   end
 
   test "should create city" do
     sign_in users(:admin)
     assert_difference('City.count') do
-      post :create, locale: :en, city: {
-        body: @berlin.body,
-        lat:  @berlin.lat,
-        lng:  @berlin.lng,
-        name: @berlin.name
+      post :create, params: {
+        locale: :en,
+        city:   {
+          body: @berlin.body,
+          lat:  @berlin.lat,
+          lng:  @berlin.lng,
+          name: @berlin.name
+        }
       }
     end
 
@@ -39,23 +42,27 @@ class CitiesControllerTest < ActionController::TestCase
   end
 
   test "should show city" do
-    get :show, locale: :en, id: @berlin
+    get :show, params: { locale: :en, id: @berlin }
     assert_response :success
   end
 
   test "should get edit" do
     sign_in users(:admin)
-    get :edit, locale: :en, id: @berlin
+    get :edit, params: { locale: :en, id: @berlin }
     assert_response :success
   end
 
   test "should update city" do
     sign_in users(:berlin_user)
-    patch :update, locale: :en, id: @berlin, city: {
-      body: @berlin.body,
-      lat:  @berlin.lat,
-      lng:  @berlin.lng,
-      name: @berlin.name
+    patch :update, params: {
+      locale: :en,
+      id:     @berlin,
+      city:   {
+        body: @berlin.body,
+        lat:  @berlin.lat,
+        lng:  @berlin.lng,
+        name: @berlin.name
+      }
     }
     assert_redirected_to city_squats_path(@berlin)
   end
@@ -63,7 +70,7 @@ class CitiesControllerTest < ActionController::TestCase
   test "should destroy city" do
     sign_in users(:admin)
     assert_difference('City.count', -1) do
-      delete :destroy, locale: :en, id: @berlin
+      delete :destroy, params: { locale: :en, id: @berlin }
     end
 
     assert_redirected_to cities_path
