@@ -19,7 +19,7 @@ class CommentsControllerTest < ActionController::TestCase
   end
 
   test "should get index" do
-    get :index, locale: :en
+    get :index, params: { locale: :en }
     assert_response :success
     assert_not_nil assigns(:comments)
     assert_includes(assigns(:comments), @comment)
@@ -28,16 +28,22 @@ class CommentsControllerTest < ActionController::TestCase
   test "should bulk_delete" do
     sign_in users(:admin)
     assert_difference('Comment.count', -2) do
-      post :bulk_delete, locale: :en, comment_ids: [comments(:berlin_comment).id, comments(:london_comment).id]
+      post :bulk_delete, params: {
+        locale: :en,
+        comment_ids: [comments(:berlin_comment).id, comments(:london_comment).id]
+      }
     end
     assert_redirected_to comments_path
   end
 
   test "should get new" do
     sign_in @not_city_user
-    get :new, locale: :en, comment: {
-      commentable_id:   @city.id,
-      commentable_type: 'City'
+    get :new, params: {
+      locale: :en,
+      comment: {
+        commentable_id:   @city.id,
+        commentable_type: 'City'
+      }
     }
     assert_response :success
   end
@@ -45,9 +51,10 @@ class CommentsControllerTest < ActionController::TestCase
   test "should create comment" do
     sign_in @not_city_user
     assert_difference('Comment.count') do
-      post :create,
-           locale:  :en,
-           comment: @valid_comment_params
+      post :create, params: {
+        locale:  :en,
+        comment: @valid_comment_params
+      }
     end
 
     assert_redirected_to comment_path(assigns(:comment))
@@ -56,9 +63,10 @@ class CommentsControllerTest < ActionController::TestCase
   test "#create should send mail" do
     sign_in @not_city_user
     assert_difference('ActionMailer::Base.deliveries.count') do
-      post :create,
-           locale:  :en,
-           comment: @valid_comment_params
+      post :create, params: {
+        locale:  :en,
+        comment: @valid_comment_params
+      }
     end
     comment_notification = ActionMailer::Base.deliveries.last
     assert_includes(comment_notification.to, @city_user.email)
@@ -66,29 +74,30 @@ class CommentsControllerTest < ActionController::TestCase
   end
 
   test "should show comment" do
-    get :show, locale: :en, id: @comment
+    get :show, params: { locale: :en, id: @comment }
     assert_response :success
   end
 
   test "should get edit" do
     sign_in users(:admin)
-    get :edit, locale: :en, id: @comment
+    get :edit, params: { locale: :en, id: @comment }
     assert_response :success
   end
 
   test "should update comment" do
     sign_in users(:admin)
-    patch :update,
-          locale:  :en,
-          id:      @comment,
-          comment: @valid_comment_params
+    patch :update, params: {
+      locale:  :en,
+      id:      @comment,
+      comment: @valid_comment_params
+    }
     assert_redirected_to comment_path(assigns(:comment))
   end
 
   test "should destroy comment" do
     sign_in users(:admin)
     assert_difference('Comment.count', -1) do
-      delete :destroy, locale: :en, id: @comment
+      delete :destroy, params: { locale: :en, id: @comment }
     end
 
     assert_redirected_to comments_path
